adding

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -8,9 +8,9 @@
  */
 
 
-int _strlen(char *s)
+size_t  _strlen(char *s)
 {
-	int length = 0;
+	size_t length = 0;
 
 	while (*s++)
 		length++;
